+++ capsicum/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased] - ReleaseDate
+
+### Added
+
+- Implemented `Send` and `Sync` for `CapChannel`.
+  ([#66](https://github.com/dlrobertson/capsicum-rs/pull/66))
+
 ## [0.3.0] - 2023-09-21
 
 ### Changed